Address @alejo0o feedback
