--- package/src/lib/components/tour-overlay/shapes/shape.types.ts
@@ -1,15 +1,0 @@
-import { LayoutRectangle, MeasureInWindowOnSuccessCallback } from "react-native";
-
-import { Motion } from "../../../SpotlightTour.context";
-
-interface RefNode {
-  measure: (callback: MeasureInWindowOnSuccessCallback) => void;
-}
-
-export interface ShapeProps {
-  motion: Motion;
-  padding: number;
-  setReference: (node?: RefNode) => void;
-  spot: LayoutRectangle;
-  useNativeDriver: boolean;
-}
